dev_requirements.txt
# Import main requirements
-r requirements.txt

django-debug-toolbar

django-factory_boy
factory_boy

# For doc
Sphinx
sphinx_rtd_theme

# For tests
coverage
unittest-xml-reporting
pylint
